Moved admin functions tests to new file
